Add "project mode" searching
When you're in a sublime project, the search starts from the root of the
project. Furthermore, multiple candidate files are sought. If there's more
than one candidate, a quick panel will show up in which the user must select
the correct candidate. From that point on it's in the cache, so the file that
the user selected will take precedence from that point on, unless Sublime is
restarted.